requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
telethon==1.34.0
anthropic==0.18.1
python-dotenv==1.0.0
pyyaml==6.0.1
